+++ src/components/icons/CloseIcon.jsx
